requirements.txt
playwright==1.53.0
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
